backend/services/ai_engine/requirements.txt
# backend/services/ai_engine/requirements.txt

# For PDF parsing
pdfminer.six

# For vector embeddings (AI matching)
sentence-transformers

# For machine learning metrics (ATS scoring)
scikit-learn

# For making HTTP requests (chatbot LLM)
requests

# For other potential uses
pymongo
python-dotenv
spacy